:recycle: [repositories] Move `hg` fixture to `tests.fixtures`
